Temporary workaround for broken AsyncImage loading
